数据库整型字段统一为INTEGER

+++ DbMonitor.Domain/Authority.cs
@@ -27,12 +27,12 @@ namespace DbMonitor.Domain
         public string AIcon { get; set; }
         public string AIconType { get; set; }
         public string AClassName { get; set; }
-        public Nullable<short> AShowNumber { get; set; }
-        public Nullable<short> AIsCommon { get; set; }
-        public Nullable<short> IsEnabled { get; set; }
-        public Nullable<int> CreatorID { get; set; }
+        public Nullable<long> AShowNumber { get; set; }
+        public Nullable<long> AIsCommon { get; set; }
+        public Nullable<long> IsEnabled { get; set; }
+        public Nullable<long> CreatorID { get; set; }
         public string CreationTime { get; set; }
-        public Nullable<int> EditorID { get; set; }
+        public Nullable<long> EditorID { get; set; }
         public string EditingTime { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
