+++ DbMonitor.Domain/User.cs
@@ -24,11 +24,11 @@ namespace DbMonitor.Domain
         public string ULoginName { get; set; }
         public string UPassword { get; set; }
         public string UName { get; set; }
-        public Nullable<short> UUserType { get; set; }
-        public Nullable<short> UIsLock { get; set; }
-        public Nullable<int> CreatorID { get; set; }
+        public Nullable<long> UUserType { get; set; }
+        public Nullable<long> UIsLock { get; set; }
+        public Nullable<long> CreatorID { get; set; }
         public string CreationTime { get; set; }
-        public Nullable<int> EditorID { get; set; }
+        public Nullable<long> EditorID { get; set; }
         public string EditingTime { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
